ACCUMULO-4558 Moved TableInfoUtil to util folder, fixed the imports to point to it
