+++ core/src/main/java/org/apache/accumulo/core/util/TableInfoUtil.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.accumulo.core.client;
+package org.apache.accumulo.core.util;
 
 import java.util.HashMap;
 import java.util.Map;
