+++ shell/src/main/java/org/apache/accumulo/shell/commands/TabletServerStatusCommand.java
@@ -19,13 +19,11 @@ package org.apache.accumulo.shell.commands;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.accumulo.core.client.AccumuloException;
 import org.apache.accumulo.core.client.TServerStatus;
 import org.apache.accumulo.core.client.admin.InstanceOperations;
 import org.apache.accumulo.shell.Shell;
 import org.apache.accumulo.shell.Shell.Command;
 import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.MissingArgumentException;
 import org.apache.commons.cli.MissingOptionException;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.Options;
@@ -44,7 +42,7 @@ public class TabletServerStatusCommand extends Command {
   }
 
   @Override
-  public int execute(final String fullCommand, final CommandLine cl, final Shell shellState) throws Exception {    
+  public int execute(final String fullCommand, final CommandLine cl, final Shell shellState) throws Exception {
     List<TServerStatus> tservers;
 
     final InstanceOperations instanceOps = shellState.getConnector().instanceOperations();
@@ -63,7 +61,7 @@ public class TabletServerStatusCommand extends Command {
     } else {
       throw new MissingOptionException("Missing options");
     }
-    
+
     if (tservers.isEmpty()) {
       throw new NotFoundException("Tablet Servers not found");
     }
@@ -87,7 +85,7 @@ public class TabletServerStatusCommand extends Command {
 
     disablePaginationOpt = new Option("np", "no-pagination", false, "disable pagination of output");
     opts.addOption(disablePaginationOpt);
-    
+
     allOption = new Option("a", "all", false, "list tablet servers status");
     opts.addOption(allOption);
